+++ packages/ui/src/accounts/modals/InsufficientFundsModal/index.ts
@@ -1,0 +1,1 @@
+export * from './InsufficientFundsModal'
